JRR[MINOR]: Added validation when the stored string is not valid.
If the string passed to the verifier is not valid (was not created by
this library) then it was crashing (with a Null Pointer exception).
If the verifier cannot be initialized correctly, it is not initialized
at all, and the verifier throws an IllegalStateException.
The string client of the library (Tomcat interface) now catches the
exception and returns false on the validation. The error is not further
pursued (if the stored credentials are not valid, no password will ever
be correct).