JRR[MINOR]: Bug fixed, crash in Tomcat if password is null.
For some reason, Tomcat calls the mutate method during login when
password is null (probably when it is not null as well). By doing it,
the method tried to geenrate a PBKDF2 from a null password resulting in
a crash while setting the starting key during the algorithm's first
cycle.
A validation was set, in which if a null password is set for mutating, a
null mutated password is returned. Tomcat worked fine with this
solution.